update i2v preprocess
